Update Phased AMM Deployment and Fair Price Discovery.md
Adding updates for details regarding the AMM Initiation Phase and AMM Balancing Phase. This includes updating to align with adjustments found in MRC09 and also to provide a note about the AMM Initiation Phase outcome that wETH exceeded MOR and thus Scenario 1 as outlined was the result to be followed.